Refactor `gs_commit_log_helper` for re-use
